requirements.txt
absl-py==1.0.0
altgraph==0.17.4
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
bleach==5.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
commonmark==0.9.1
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
docutils==0.18.1
entrypoints==0.4
executing==0.8.3
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.11.0
ipyplot==1.1.1
ipython==8.2.0
jedi==0.18.1
jeepney==0.8.0
jupyter-client==7.2.1
jupyter-core==4.9.2
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
matplot==0.1.9
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.8.2
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
Pygments==2.11.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyloco==0.0.139
pyparsing==3.0.8
python-dateutil==2.8.2
pyzmq==22.3.0
readme-renderer==35.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.1
rsa==4.8
SecretStorage==3.3.2
shortuuid==1.0.8
SimpleWebSocketServer==0.1.1
six==1.16.0
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tornado==6.1
traitlets==5.1.1
twine==4.0.0
typing-extensions==4.1.1
urllib3==1.26.9
ushlex==0.99.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.0
wrapt==1.14.0
zipp==3.7.0
